+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Hackpad-cli changelog
 v0.1.0 - wip
 ------------
 
+- add a `stats` comand to show how much pads are cached
+- move `-r` (refresh) and `-u` (urls) options to specific method options
 - add a User-Agent so that hackpad knows what is poking them
 - completed test coverage
 - prevent trailing slash on site config param to fuck up urls